backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.1.0
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
uagents>=0.13.0
